Made separate folder for tests
